+++ CHANGELOG.md
@@ -4,7 +4,7 @@ SSLkeylog
 0.2.0
 -----
 
-(unreleased)
+2015-04-14
 
 Backwards incompatible feature release.
 
@@ -13,7 +13,13 @@ Backwards incompatible feature release.
 
   - +Break
     The library no longer extends (monkeypatches) the `OpenSSL::SSL::SSLSocket`
-    class.
+    class. This functionality is now provided by module functions on
+    `SSLkeylog::OpenSSL`.
+
+  - +New
+    A simple tracing and logging system built on top of the standard `Logging`
+    library and the Ruby 2.0 `TracePoint` API that allows SSL client
+    connections to be logged.
 
 
 0.1.0